+++ src/displayBuffer.h
@@ -17,7 +17,7 @@ void fillBufferWithValue(displayBuffer* buffer, uint8_t value);
 void changeBrightness(displayBuffer* buffer, uint8_t intensity);
 
 void addDisplayBufferToDisplayBuffer(displayBuffer* sourceBuffer, displayBuffer* inputBuffer, uint8_t x, uint8_t y);
-void addUint8TBufferToDisplay(displayBuffer* sourceBuffer, const uint8_t* inputBuffer, uint8_t w, uint8_t h, uint8_t x, uint8_t y);
+void addUint8TBufferToDisplay(displayBuffer* sourceBuffer, const uint8_t* inputBuffer, uint16_t w, uint16_t h, uint8_t x, uint8_t y);
 
 uint8_t convertPixelsToOledPixel(uint8_t p1, uint8_t p2);
 void convertNormalDisplayBufferToOledBuffer(displayBuffer* buffer);
